requirements.txt
langchain
langgraph
chromadb
sentence-transformers
streamlit
python-dotenv
langchain-groq